Move common classes from client-device-connection-infinispan module into client-device-connection-base module. Unit tests run.
