+++ client-device-connection-base/src/main/java/org/eclipse/hono/deviceconnection/common/CommonCacheConfig.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.deviceconnection.infinispan.client;
+package org.eclipse.hono.deviceconnection.common;
 
 import com.google.common.base.MoreObjects;
