+++ client-device-connection-base/src/main/java/org/eclipse/hono/deviceconnection/common/UnknownStatusProvider.java
@@ -12,7 +12,7 @@
  */
 
 
-package org.eclipse.hono.deviceconnection.infinispan.client;
+package org.eclipse.hono.deviceconnection.common;
 
 import java.util.Collection;
 import java.util.Set;
