+++ client-device-connection-base/src/test/resources/common-cache-options.yaml
