+++ services/command-router/src/main/java/org/eclipse/hono/commandrouter/AdapterInstanceStatusService.java
@@ -13,7 +13,7 @@
 
 package org.eclipse.hono.commandrouter;
 
-import org.eclipse.hono.deviceconnection.infinispan.client.AdapterInstanceStatusProvider;
+import org.eclipse.hono.deviceconnection.common.AdapterInstanceStatusProvider;
 import org.eclipse.hono.util.Lifecycle;
 
 /**
